third_party/blink/web_tests/platform/generic/http/tests/cache/subresource-multiple-instances-expected.txt

Verifies that resources are loaded only once during the initial document load, but are reloaded if they are requested again later and can not be cached.

PASS firstRandomNumber is non-zero.
PASS secondRandomNumber is non-zero.
PASS thirdRandomNumber is non-zero.
PASS firstRandomNumber == secondRandomNumber is true
PASS firstRandomNumber != thirdRandomNumber is true
PASS successfullyParsed is true

TEST COMPLETE
